Generates SPM Packages requiring Swift 5
This updates create_packages.py (and parallel code in the
create_packages Python notebook) so that generated Package.swift files
have a swift-tools-version of 5.0 instead of 4.2.

This implicitly requires a Swift language version of 5.0 in order to
build the package. This is what we want, since we depend on a recent
release of Swift with Tensorflow and all such releases variants of
Swift 5.

The main practical benefit of this change is that if you generate an
Xcode project from one of these packages, the Xcode project will
understand the required language version, so you will not need to set
that manually. And it might help you notice quicker when you haven't
enabled the Swift with Tensorflow toolchain are using Swift 4.2 by
mistake.